requirements.txt
streamlit==1.28
google-generativeai
textblob
pandas
matplotlib
pyngrok
python-dotenv
seaborn